build!: draft on AbstractGraph
